refactor: chat types are now in chat folder
